# server/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Bzz System API",
    "description": "Bzz system API for controlling cars",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api/",
      "description": "From localhost api"
    }
  ],
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/model": {
      "get": {
        "security": [
          {
             "bearerAuth": []
          }
        ],
        "tags": [
          "Model"
        ],
        "summary": "Get all models in system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Model"
            }
          }
        }
      },
      "post": {
        "security": [
          {
             "bearerAuth": []
          }
        ],
        "tags": [
          "Model"
        ],
        "summary": "Create a new model in system",
        "requestBody": {
          "description": "Model Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/Model"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/Model"
            }
          },
          "400": {
            "description": "Failed. Bad post data."
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Model": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "prices": {
          "$ref": "#/definitions/Price"
        }
      }
    },
...
    "Price": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "minute": {
          "type": "number"
        },
        "hour": {
          "type": "number"
        },
        "day": {
          "type": "number"
        },
        "week": {
          "type": "number"
        },
        "month": {
          "type": "number"
        },
        "km": {
          "type": "number"
        }
      }
    }
  },
  "components": {        
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}